Added module for Priority Encoder
It will generate DACK to I/O devices based on the DREQ and some other inputs like -
SenseDreq which is Command[6] register.
SenseDack which is Command[7] register.
RotatingPriority is Command[4] register.
Mask is 4 bits, tells which channels need to be ignored.

Generates ReqID, it tells which channel is being serviced.
ValidReqID tells if ReqID is valid or not.